% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{get_download_cache}
\alias{get_download_cache}
\title{Get download cache directory}
\usage{
get_download_cache(create = TRUE)
}
\arguments{
\item{create}{Logical - Whether to create the directory if it doesn't exist
(default: TRUE)}
}
\value{
Character - Path to the cache directory
}
\description{
Returns the path to the default GAEZ download directory, creating it if
necessary.
}
\examples{
\dontrun{
cache_dir <- get_download_cache()
print(paste("GAEZ files stored in:", cache_dir))
}

}
