% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_gaez_scenarios}
\alias{get_gaez_scenarios}
\title{Get GAEZ v5 Scenarios Table}
\usage{
get_gaez_scenarios()
}
\value{
A tibble with multiple types of scenario information including:
\itemize{
\item Time periods (HP8100, HP0120, FP2140, FP4160, FP6180, FP8100)
\item Climate models (AGERA5, ENSEMBLE, GFDL-ESM4, etc.)
\item SSP scenarios (HIST, SSP126, SSP370, SSP585)
\item Water management codes (HRLM, HILM, LRLM, LILM)
\item Water supply codes (WSI, WSR, WST)
}
}
\description{
Returns reference data for time periods, climate models, SSP scenarios, and
water management codes used in GAEZ v5.
}
\examples{
\dontrun{
# Get all scenarios
scenarios <- get_gaez_scenarios()
View(scenarios)

# Get just time periods
library(dplyr)
time_periods <- get_gaez_scenarios() \%>\%
  filter(!is.na(time_period))
}

}
