% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.gaez_batch_result}
\alias{print.gaez_batch_result}
\title{Print method for batch download results}
\usage{
\method{print}{gaez_batch_result}(x, ...)
}
\arguments{
\item{x}{A gaez_batch_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Provides clear summary of batch download operations
}
