% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_gaez_examples}
\alias{show_gaez_examples}
\title{Show GAEZ v5 function examples}
\usage{
show_gaez_examples()
}
\value{
Invisible NULL. Prints examples to console.
}
\description{
Displays example usage for the main functions in the gaezv5 package.
Helpful for learning the package workflow and discovering functionality.
}
\examples{
show_gaez_examples()

}
