% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaezv5-package.R
\docType{package}
\name{gaezv5-package}
\alias{gaezv5-package}
\alias{gaezv5}
\title{gaezv5: Download and Process GAEZ v5 Agricultural Data}
\description{
Tools for downloading and working with Global Agro-Ecological Zones (GAEZ) version 5 data from the Food and Agriculture Organization (FAO). GAEZ v5 provides comprehensive global data on agricultural potential, crop suitability, attainable yields, and agro-climatic resources. This package simplifies data access through intuitive functions for searching variables, building download URLs, and batch downloading datasets across multiple crops, scenarios, and time periods. Features efficient parallel downloads, time series support, and flexible parameter handling for climate change impact studies.

The gaezv5 package provides tools for downloading and working with GAEZ
(Global Agro-Ecological Zones) version 5 data from the FAO. GAEZ v5 provides
comprehensive global data on agricultural potential, crop suitability, yield
estimates, and agro-climatic resources.
}
\section{Main functions}{

\itemize{
\item \code{\link{lookup_gaez_variable}}: Find variable codes
\item \code{\link{lookup_gaez_crop}}: Find crop codes
\item \code{\link{build_gaez_url}}: Construct download URLs
\item \code{\link{download_gaez_dataset}}: Download single dataset
\item \code{\link{batch_download_gaez_datasets}}: Download multiple datasets
\item \code{\link{list_gaez_crops}}: List available crops
\item \code{\link{list_downloaded_files}}: Inventory of local files
\item \code{\link{check_url_exists}}: Validate URLs before download
}
}

\section{Data objects}{

\itemize{
\item \code{\link{gaez_variables}}: Table of all GAEZ variables
\item \code{\link{gaez_crops}}: Table of all crop codes
\item \code{\link{gaez_scenarios}}: Time periods, climate models, SSPs
\item \code{\link{gaez_url_structure}}: URL construction patterns
}
}

\section{Getting started}{

To get started with gaezv5:

\preformatted{
# 1. Browse available crops
list_gaez_crops()

# 2. Download a dataset
result <- download_gaez_dataset(
  crop = "maize",
  time_period = "HP0120",
  climate_model = "AGERA5"
)

# 3. Load the raster
library(terra)
r <- rast(result$file_path)
plot(r)
}
}

\section{GAEZ v5 Themes}{

GAEZ v5 data is organized into 6 themes:
\enumerate{
\item Land and Water Resources
\item Agro-climatic Resources
\item Agro-climatic Potential Yield
\item Suitability and Attainable Yield
\item Actual Yields and Production
\item Yield and Production Gaps
}
}

\section{Learn more}{

\itemize{
\item GAEZ v5 portal: \url{https://gaez.fao.org/}
\item Package vignettes: \code{browseVignettes("gaezv5")}
\item Examples: \code{show_gaez_examples()}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpwjoseph/gaezv5}
  \item Report bugs at \url{https://github.com/jpwjoseph/gaezv5/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/jpwjoseph/gaezv5}
  \item Report bugs at \url{https://github.com/jpwjoseph/gaezv5/issues}
}

}
\author{
\strong{Maintainer}: Julian Joseph \email{joseph@iiasa.ac.at} (\href{https://orcid.org/0000-0002-3844-7807}{ORCID})

Other contributors:
\itemize{
  \item IIASA [copyright holder]
}


Julian Joseph \email{joseph@iiasa.ac.at}
}
\keyword{internal}
